+++ src/tokens/RankToken.sol
@@ -8,8 +8,7 @@ import "@peeramid-labs/eds/src/abstracts/ERC7746Middleware.sol";
 import "@peeramid-labs/eds/src/libraries/LibMiddleware.sol";
 import {IERC1155} from "@openzeppelin/contracts/interfaces/IERC1155.sol";
 import {IERC165} from "@openzeppelin/contracts/interfaces/IERC165.sol";
-import {IFellowship} from "../interfaces/IFellowship.sol";
-import "@peeramid-labs/eds/src/abstracts/InstallerCloneable.sol";
+import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 //ToDo: it was planned to make it track for highest token users hold (their rank), right now it's not implemented. Yet.
 
 /**
@@ -17,7 +16,7 @@ import "@peeramid-labs/eds/src/abstracts/InstallerCloneable.sol";
  * @author Peersky
  * @notice RankToken is a composite ERC1155 token that is used to track user ranks
  */
-contract RankToken is InstallerClonable, IFellowship, LockableERC1155, IRankToken, ERC7746Middleware {
+contract RankToken is LockableERC1155, IRankToken, ERC7746Middleware, OwnableUpgradeable {
     struct Storage {
         string _contractURI;
     }
@@ -31,19 +30,17 @@ contract RankToken is InstallerClonable, IFellowship, LockableERC1155, IRankToke
         }
     }
 
-    constructor(
+    constructor(string memory uri_, string memory cURI, address accessLayer, address owner_) {
+        initialize(uri_, cURI, accessLayer, owner_);
+    }
+
+    function initialize(
         string memory uri_,
         string memory cURI,
         address accessLayer,
-        address token,
-        uint256 defaultCost,
-        address defaultAdmin
-    ) TokenizedDistributor(token, defaultCost, defaultAdmin) {
-        initialize(uri_, cURI, accessLayer, token, defaultCost, defaultAdmin);
-    }
-
-    function initialize(string memory uri_, string memory cURI, address accessLayer, address token, uint256 defaultCost, address defaultAdmin) public initializer {
-        // __Ownable_init(owner_);
+        address owner_
+    ) public initializer {
+        __Ownable_init(owner_);
         _setURI(uri_);
         getStorage()._contractURI = cURI;
         LibMiddleware.LayerStruct[] memory layers = new LibMiddleware.LayerStruct[](1);
@@ -59,14 +56,6 @@ contract RankToken is InstallerClonable, IFellowship, LockableERC1155, IRankToke
 
     function contractURI() public view returns (string memory) {
         return getStorage()._contractURI;
-    }
-
-    function setURI(string memory uri_) public ERC7746C(msg.sig, msg.sender, msg.data, 0) {
-        _setURI(uri_);
-    }
-
-    function setContractURI(string memory uri_) public ERC7746C(msg.sig, msg.sender, msg.data, 0) {
-        getStorage()._contractURI = uri_;
     }
 
     function _mintRank(address to, uint256 amount, uint256 level, bytes memory data) private {
@@ -141,105 +130,11 @@ contract RankToken is InstallerClonable, IFellowship, LockableERC1155, IRankToke
         super.burn(account, id, value);
     }
 
-    /**
-     * @dev Internal function to create a new game with the specified parameters
-     * @param params Struct containing all necessary parameters for game creation
-     * @notice This function handles the core game creation logic, including:
-     *         - Setting up the game state
-     *         - Configuring the coin vending system
-     *         - Emitting the game creation event
-     */
-    function createThread(LibRankify.NewGameParams memory params) private nonReentrant {
-        //TODO: add this back in start  game to verify commitment from game master
-        //  bytes32 digest = _hashTypedDataV4(
-        //     keccak256(
-        //         abi.encode(
-        //             keccak256(
-        //                 "AttestGameCreation(uint256 gameId,uint256 commitment)"
-        //             ),
-        //             params.gameId,
-        //             params.gmCommitment
-        //         )
-        //     )
-        // );
-
-        LibRankify.newGame(params);
-        LibCoinVending.ConfigPosition memory emptyConfig;
-        LibCoinVending.configure(bytes32(params.gameId), emptyConfig);
-        emit gameCreated(params.gameId, params.gameMaster, msg.sender, params.gameRank);
+    function setContractURI(string memory uri_) public onlyOwner {
+        getStorage()._contractURI = uri_;
     }
 
-    /**
-     * @dev External function to create a new game
-     * @param params Input parameters for creating a new game
-     * @notice This function:
-     *         - Validates the contract is initialized
-     *         - Processes input parameters
-     *         - Creates a new game with specified settings
-     * @custom:security nonReentrant
-     */
-    function instantiate(bytes32 id, bytes calldata args) public nonReentrant {
-        IRankifyInstance.NewGameParamsInput memory params = abi.decode(args, (IRankifyInstance.NewGameParamsInput));
-        // function instantiate() public {
-        LibRankify.enforceIsInitialized();
-        LibRankify.InstanceState storage settings = LibRankify.instanceState();
-        LibRankify.NewGameParams memory newGameParams = LibRankify.NewGameParams({
-            gameId: settings.numGames + 1,
-            gameRank: params.gameRank,
-            creator: msg.sender,
-            minPlayerCnt: params.minPlayerCnt,
-            maxPlayerCnt: params.maxPlayerCnt,
-            gameMaster: params.gameMaster,
-            nTurns: params.nTurns,
-            voteCredits: params.voteCredits,
-            minGameTime: params.minGameTime,
-            timePerTurn: params.timePerTurn,
-            timeToJoin: params.timeToJoin,
-            metadata: params.metadata
-        });
-
-        createGame(newGameParams);
-    }
-
-    function install(
-        IDistributor distributor,
-        bytes32 distributionId,
-        bytes calldata args
-    ) external payable returns (uint256 instanceId)
-    {
-      return super._install(distributor, distributionId, args);
-    }
-
-
-
-    /**
-     * @dev Returns the current state of the contract
-     * @return LibRankify.InstanceState The current state of the contract
-     */
-    function getContractState() public pure returns (LibRankify.InstanceState memory) {
-        LibRankify.InstanceState memory state = LibRankify.instanceState();
-        return state;
-    }
-
-    /**
-     * @dev Estimates the price of a game with the specified minimum game time
-     * @param minGameTime The minimum game time
-     * @return uint256 The estimated price of the game
-     */
-    function estimateThreadPrice(uint128 minGameTime) public pure returns (uint256) {
-        LibRankify.InstanceState memory state = LibRankify.instanceState();
-        return LibRankify.getGamePrice(minGameTime, state.commonParams);
-    }
-
-    function exitRankToken(uint256 rankId, uint256 amount) external {
-        require(amount != 0, "cannot specify zero exit amount");
-        LibRankify.InstanceState storage state = LibRankify.instanceState();
-        LibRankify.CommonParams storage commons = state.commonParams;
-        IRankToken rankContract = IRankToken(commons.rankTokenAddress);
-        DistributableGovernanceERC20 tokenContract = DistributableGovernanceERC20(commons.derivedToken);
-        uint256 _toMint = amount * (commons.principalCost * (commons.minimumParticipantsInCircle ** rankId));
-        rankContract.burn(msg.sender, rankId, amount);
-        tokenContract.mint(msg.sender, _toMint);
-        emit RankTokenExited(msg.sender, rankId, amount, _toMint);
+    function setURI(string memory uri_) public onlyOwner {
+        _setURI(uri_);
     }
 }