+++ test/MAODistribution.ts
@@ -3,7 +3,7 @@
 import { deployments, ethers, getNamedAccounts } from 'hardhat';
 import hre from 'hardhat';
 import { expect } from 'chai';
-import { MAODistribution, DAODistributor, Rankify, RankifyDiamondInstance } from '../types';
+import { MAODistribution, DAODistributor, Rankify, RankifyDiamondInstance, RankToken } from '../types';
 import { setupTest } from './utils';
 
 import { getCodeIdFromArtifact } from '../scripts/getCodeId';
@@ -21,6 +21,7 @@ describe('MAODistribution', async function () {
   beforeEach(async function () {
     const setup = await setupTest();
     addr = setup.adr;
+    console.log('setup complete');
     contract = setup.env.maoDistribution;
     const maoCode = await hre.ethers.provider.getCode(contract.address);
     maoId = ethers.utils.keccak256(maoCode);
@@ -53,6 +54,8 @@ describe('MAODistribution', async function () {
       });
     });
     it('Can instantiate a distribution', async () => {
+      const { owner } = await getNamedAccounts();
+      const oSigner = await ethers.getSigner(owner);
       const distributorArguments: MAODistribution.DistributorArgumentsStruct = {
         tokenSettings: {
           tokenName: 'tokenName',
@@ -63,14 +66,13 @@ describe('MAODistribution', async function () {
           rankTokenURI: 'https://example.com/rank',
           principalCost: 1,
           principalTimeConstant: 1,
+          owner: oSigner.address,
         },
       };
       // Encode the arguments using generateDistributorData
       const data = generateDistributorData(distributorArguments);
-
       const token = await deployments.get('Rankify');
-      const { owner } = await getNamedAccounts();
-      const oSigner = await ethers.getSigner(owner);
+
       const tokenContract = new ethers.Contract(token.address, token.abi, oSigner) as Rankify;
       await tokenContract.mint(oSigner.address, ethers.utils.parseUnits('100', 9));
       await tokenContract.approve(distributorContract.address, ethers.constants.MaxUint256);
@@ -87,6 +89,44 @@ describe('MAODistribution', async function () {
         evts[0].args.instances[2],
       )) as RankifyDiamondInstance;
       expect((await ACIDContract.functions['getGM(uint256)'](0))[0]).to.equal(ethers.constants.AddressZero);
+    });
+    it('Can allows initiator to set contractURI', async () => {
+      const { owner } = await getNamedAccounts();
+      const oSigner = await ethers.getSigner(owner);
+      const distributorArguments: MAODistribution.DistributorArgumentsStruct = {
+        tokenSettings: {
+          tokenName: 'tokenName',
+          tokenSymbol: 'tokenSymbol',
+        },
+        rankifySettings: {
+          rankTokenContractURI: 'https://example.com/rank',
+          rankTokenURI: 'https://example.com/rank',
+          principalCost: 1,
+          principalTimeConstant: 1,
+          owner,
+        },
+      };
+      // Encode the arguments using generateDistributorData
+      const data = generateDistributorData(distributorArguments);
+
+      const token = await deployments.get('Rankify');
+      const tokenContract = new ethers.Contract(token.address, token.abi, oSigner) as Rankify;
+      await tokenContract.mint(oSigner.address, ethers.utils.parseUnits('100', 9));
+      await tokenContract.approve(distributorContract.address, ethers.constants.MaxUint256);
+
+      const tx = await distributorContract.connect(oSigner).instantiate(distrId, data);
+      await expect(tx).not.reverted;
+      expect((await distributorContract.functions.getDistributions()).length).to.equal(1);
+      const filter = distributorContract.filters.Instantiated();
+      const evts = await distributorContract.queryFilter(filter);
+      expect(evts.length).to.equal(1);
+
+      const RankTokenContract = (await ethers.getContractAt('RankToken', evts[0].args.instances[11])) as RankToken;
+      await RankTokenContract.connect(oSigner).setContractURI('foo');
+      await RankTokenContract.connect(oSigner).setURI('Uri_foo');
+      expect(await RankTokenContract.contractURI()).to.equal('foo');
+      expect(await RankTokenContract.uri(1)).to.equal('Uri_foo');
+      expect(await RankTokenContract.owner()).to.be.equal(oSigner.address);
     });
   });
 });